refactor(select): rename source files to follow convention
All `[ng-model]` directive source files under `lib/directive` have a
name of the form `ng_model*.dart`. Rename:
- `input_select.dart` to `ng_model_select.dart`.
- `input_select_spec.dart` to `ng_model_select_spec.dart`.
